Translated File
Used ChatGPT to translate all of the comments from the original file to English while leaving the rest of the file in tact. I will be starting with this file for my original build. Comparing them side to side in VS code everything looked good. Variables were translated as well and I will need to test them going forward.